+++ src/main/java/co/edu/uniquindio/demo/demo/ProductoPerecedero.java
@@ -6,7 +6,7 @@ public class ProductoPerecedero extends Producto{
 
     private String fechaVencimiento;
 
-    public ProductoPerecedero(String codigo, String nombre, String descripcion, double valorUnitario, int cantidadExistencia, String fechaVencimiento) {
+    public ProductoPerecedero(String codigo, String nombre, String descripcion, Float valorUnitario, int cantidadExistencia, String fechaVencimiento) {
         super(codigo, nombre, descripcion, valorUnitario, cantidadExistencia);
         this.fechaVencimiento = fechaVencimiento;
     }
